+++ source/main/commands/import.js
@@ -30,7 +30,7 @@ class ImportFiles extends ZettlrCommand {
     */
   async run (evt, arg) {
     if (!this._app.getCurrentDir()) {
-      global.ipc.notify(trans('system.import_no_directory'))
+      global.notify.normal(trans('system.import_no_directory'))
       return false
     }
 
@@ -54,25 +54,25 @@ class ImportFiles extends ZettlrCommand {
     }
 
     // Now import.
-    global.ipc.notify(trans('system.import_status'))
+    global.notify.normal(trans('system.import_status'))
     try {
       let ret = await makeImport(fileList, this._app.getCurrentDir(), (file, error) => {
         // This callback gets called whenever there is an error while running pandoc.
-        global.ipc.notify(trans('system.import_error', path.basename(file)))
+        global.notify.normal(trans('system.import_error', path.basename(file)))
       }, (file) => {
         // And this on each success!
-        global.ipc.notify(trans('system.import_success', path.basename(file)))
+        global.notify.normal(trans('system.import_success', path.basename(file)))
       })
 
       if (ret.length > 0) {
         // Some files failed to import.
-        global.ipc.notify(trans('system.import_fail', ret.length, ret.map((x) => { return path.basename(x) }).join(', ')))
+        global.notify.normal(trans('system.import_fail', ret.length, ret.map((x) => { return path.basename(x) }).join(', ')))
       }
     } catch (e) {
       // There has been an error on importing (e.g. Pandoc was not found)
       // This catches this and displays it.
       global.log.error(e.message, e)
-      global.ipc.notify(e.message)
+      global.notify.normal(e.message)
     }
   }
 }