+++ source/main/commands/print.js
@@ -56,7 +56,7 @@ class Print extends ZettlrCommand {
         // Now we'll need to open the print window.
         this._app.showPrintWindow(target)
       })
-      .catch((err) => { global.ipc.notify(err.name + ': ' + err.message) }) // Error may be thrown
+      .catch((err) => { global.notify.normal(err.name + ': ' + err.message) }) // Error may be thrown
   }
 }